LINQ: split each method groups into a separate file
