+++ Doxense.Core/Linq/Async/AsyncQuery.EmptyQuery.cs
@@ -204,89 +204,6 @@ namespace SnowBank.Linq
 
 		}
 
-		/// <summary>A query that will return only a single element</summary>
-		internal sealed class SingletonQuery<TElement> : AsyncLinqIterator<TElement>
-		{
-
-			private Delegate? Factory { get; }
-
-			private bool Called { get; set; }
-
-			private SingletonQuery(Delegate lambda, CancellationToken ct)
-			{
-				Contract.Debug.Requires(lambda != null);
-				this.Factory = lambda;
-				this.Cancellation = ct;
-			}
-
-			public SingletonQuery(Func<TElement> lambda, CancellationToken ct)
-				: this((Delegate) lambda, ct)
-			{ }
-
-			public SingletonQuery(Func<Task<TElement>> lambda, CancellationToken ct)
-				: this((Delegate) lambda, ct)
-			{ }
-
-			public SingletonQuery(Func<ValueTask<TElement>> lambda, CancellationToken ct)
-				: this((Delegate) lambda, ct)
-			{ }
-
-			public SingletonQuery(Func<CancellationToken, Task<TElement>> lambda, CancellationToken ct)
-				: this((Delegate) lambda, ct)
-			{ }
-
-			public SingletonQuery(Func<CancellationToken, ValueTask<TElement>> lambda, CancellationToken ct)
-				: this((Delegate) lambda, ct)
-			{ }
-
-			public override CancellationToken Cancellation { get; }
-
-			/// <inheritdoc />
-			protected override SingletonQuery<TElement> Clone() => new(this.Factory!, this.Cancellation);
-
-			/// <inheritdoc />
-			protected override ValueTask<bool> OnFirstAsync()
-			{
-				if (this.Factory == null)
-				{
-					return new(false);
-				}
-
-				this.Called = false;
-				return new(true);
-			}
-
-			/// <inheritdoc />
-			protected override async ValueTask<bool> OnNextAsync()
-			{
-				var factory = this.Factory;
-				if (factory == null || this.Called)
-				{
-					return await this.Completed().ConfigureAwait(false);
-				}
-
-				this.Called = true;
-				var current = factory switch
-				{
-					Func<TElement> fn => fn(),
-					Func<Task<TElement>> fn => await fn().ConfigureAwait(false),
-					Func<CancellationToken, Task<TElement>> fn => await fn(this.Cancellation).ConfigureAwait(false),
-					Func<ValueTask<TElement>> fn => await fn().ConfigureAwait(false),
-					Func<CancellationToken, ValueTask<TElement>> fn => await fn(this.Cancellation).ConfigureAwait(false),
-					_ => throw new InvalidOperationException("Unsupported delegate type")
-				};
-
-				return Publish(current);
-			}
-
-			/// <inheritdoc />
-			protected override ValueTask Cleanup()
-			{
-				return default;
-			}
-
-		}
-
 	}
 
 }
