+++ Doxense.Core/Linq/Async/AsyncQuery.Iterators.cs
@@ -455,4 +455,5 @@ namespace SnowBank.Linq
 		#endregion
 
 	}
+
 }
