+++ Doxense.Core/Linq/Async/AsyncQuery.cs
@@ -55,7 +55,7 @@ namespace SnowBank.Linq
 		public static IAsyncLinqQuery<T> Singleton<T>(T value, CancellationToken ct = default)
 		{
 			//note: we can't call this method Single<T>(T), because then Single<T>(Func<T>) would be ambiguous with Single<Func<T>>(T)
-			return new EnumerableSequence<T>([ value ], ct);
+			return new EnumerableAsyncQuery<T>([ value ], ct);
 		}
 
 		/// <summary>Returns an async sequence which will produce a single element, using the specified lambda</summary>
@@ -94,7 +94,7 @@ namespace SnowBank.Linq
 		{
 			Contract.NotNull(source);
 
-			return new EnumerableSequence<T>(source, ct);
+			return new EnumerableAsyncQuery<T>(source, ct);
 		}
 
 		/// <summary>Wraps an async lambda into an async sequence that will return the result of the lambda</summary>
@@ -651,7 +651,7 @@ namespace SnowBank.Linq
 			Contract.NotNull(source);
 			Contract.NotNull(keySelector);
 
-			return new OrderedSequence<TSource, TKey>(source, keySelector, comparer ?? Comparer<TKey>.Default, descending: false, parent: null);
+			return new OrderedAsyncQuery<TSource, TKey>(source, keySelector, comparer ?? Comparer<TKey>.Default, descending: false, parent: null);
 		}
 
 		[Pure, LinqTunnel]
@@ -660,7 +660,7 @@ namespace SnowBank.Linq
 			Contract.NotNull(source);
 			Contract.NotNull(keySelector);
 
-			return new OrderedSequence<TSource, TKey>(source, keySelector, comparer ?? Comparer<TKey>.Default, descending: true, parent: null);
+			return new OrderedAsyncQuery<TSource, TKey>(source, keySelector, comparer ?? Comparer<TKey>.Default, descending: true, parent: null);
 		}
 
 		[Pure, LinqTunnel]
@@ -870,590 +870,6 @@ namespace SnowBank.Linq
 
 		#endregion
 
-		#region Aggregate...
-
-		/// <summary>Applies an accumulator function over an async sequence.</summary>
-		public static async Task<TSource> AggregateAsync<TSource>(this IAsyncQuery<TSource> source, [InstantHandle] Func<TSource, TSource, TSource> aggregator)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(aggregator);
-
-			await using (var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All))
-			{
-				Contract.Debug.Assert(iterator != null, "The sequence returned a null async iterator");
-
-				if (!(await iterator.MoveNextAsync().ConfigureAwait(false)))
-				{
-					throw ErrorNoElements();
-				}
-
-				var item = iterator.Current;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					item = aggregator(item, iterator.Current);
-				}
-
-				return item;
-			}
-		}
-
-		/// <summary>Applies an accumulator function over an async sequence.</summary>
-		public static async Task<TAggregate> AggregateAsync<TSource, TAggregate>(this IAsyncQuery<TSource> source, TAggregate seed, [InstantHandle] Func<TAggregate, TSource, TAggregate> aggregator)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(aggregator);
-
-			//TODO: optimize this to not have to allocate lambdas!
-			var accumulate = seed;
-			await ForEachAsync(source, (x) => { accumulate = aggregator(accumulate, x); }).ConfigureAwait(false);
-			return accumulate;
-		}
-
-		/// <summary>Applies an accumulator function over an async sequence.</summary>
-		public static async Task<TAggregate> AggregateAsync<TSource, TAggregate>(this IAsyncQuery<TSource> source, TAggregate seed, [InstantHandle] Action<TAggregate, TSource> aggregator)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(aggregator);
-
-			await ForEachAsync(
-				source,
-				(Fn: aggregator, Acc: seed),
-				static (s, x) => s.Fn(s.Acc, x)
-			).ConfigureAwait(false);
-
-			return seed;
-		}
-
-		/// <summary>Applies an accumulator function over an async sequence.</summary>
-		public static async Task<TResult> AggregateAsync<TSource, TAggregate, TResult>(this IAsyncQuery<TSource> source, TAggregate seed, [InstantHandle] Func<TAggregate, TSource, TAggregate> aggregator, [InstantHandle] Func<TAggregate, TResult> resultSelector)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(aggregator);
-			Contract.NotNull(resultSelector);
-
-			var accumulate = seed;
-			await ForEachAsync(source, (x) => { accumulate = aggregator(accumulate, x); }).ConfigureAwait(false);
-			return resultSelector(accumulate);
-		}
-
-		/// <summary>Applies an accumulator function over an async sequence.</summary>
-		public static async Task<TResult> AggregateAsync<TSource, TAggregate, TResult>(this IAsyncQuery<TSource> source, TAggregate seed, [InstantHandle] Action<TAggregate, TSource> aggregator, [InstantHandle] Func<TAggregate, TResult> resultSelector)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(aggregator);
-			Contract.NotNull(resultSelector);
-
-			await ForEachAsync(
-				source,
-				(Fn: aggregator, Acc: seed),
-				static (s, x) => s.Fn(s.Acc, x)
-			).ConfigureAwait(false);
-
-			return resultSelector(seed);
-		}
-
-		#endregion
-
-		#region First...
-
-		/// <summary>Returns the first element of an async sequence, or an exception if it is empty</summary>
-		public static Task<T> FirstAsync<T>(this IAsyncQuery<T> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.FirstAsync();
-			}
-
-			return Impl(source);
-
-			static async Task<T> Impl(IAsyncQuery<T> source)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Head);
-
-				if (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					return iterator.Current;
-				}
-
-				throw ErrorNoElements();
-			}
-		}
-
-		/// <summary>Returns the first element of an async sequence, or an exception if it is empty</summary>
-		public static Task<T> FirstAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.FirstAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, bool> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Iterator);
-
-				while(await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (predicate(iterator.Current))
-					{
-						return iterator.Current;
-					}
-				}
-
-				throw ErrorNoElements();
-			}
-		}
-
-		/// <summary>Returns the first element of an async sequence, or an exception if it is empty</summary>
-		public static Task<T> FirstAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.FirstAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Iterator);
-
-				var ct = source.Cancellation;
-				while(await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (await predicate(iterator.Current, ct).ConfigureAwait(false))
-					{
-						return iterator.Current;
-					}
-				}
-
-				throw ErrorNoElements();
-			}
-		}
-
-		#endregion
-
-		#region FirstOrDefaultAsync...
-
-		/// <summary>Returns the first element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T?> FirstOrDefaultAsync<T>(this IAsyncQuery<T> source)
-			=> FirstOrDefaultAsync(source, default(T?));
-
-		/// <summary>Returns the first element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T> FirstOrDefaultAsync<T>(this IAsyncQuery<T> source, T defaultValue)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.FirstOrDefaultAsync(defaultValue);
-			}
-
-			return Impl(source, defaultValue);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, T defaultValue)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Head);
-
-				if (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					return iterator.Current;
-				}
-
-				return defaultValue;
-			}
-		}
-
-		/// <summary>Returns the first element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T?> FirstOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-			=> FirstOrDefaultAsync(source, predicate!, default(T?));
-
-		/// <summary>Returns the first element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T> FirstOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate, T defaultValue)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.FirstOrDefaultAsync(predicate, defaultValue);
-			}
-
-			return Impl(source, predicate, defaultValue);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, bool> predicate, T defaultValue)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Iterator);
-
-				while(await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (predicate(iterator.Current))
-					{
-						return iterator.Current;
-					}
-				}
-
-				return defaultValue;
-			}
-		}
-
-		/// <summary>Returns the first element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T?> FirstOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-			=> FirstOrDefaultAsync(source, predicate!, default(T?));
-
-		/// <summary>Returns the first element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T> FirstOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate, T defaultValue)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.FirstOrDefaultAsync(predicate, defaultValue);
-			}
-
-			return Impl(source, predicate, defaultValue);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate, T defaultValue)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Iterator);
-
-				var ct = source.Cancellation;
-				while(await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (await predicate(iterator.Current,ct).ConfigureAwait(false))
-					{
-						return iterator.Current;
-					}
-				}
-
-				return defaultValue;
-			}
-		}
-
-		#endregion
-
-		#region SingleAsync...
-
-		/// <summary>Returns the first and only element of an async sequence, or an exception if it is empty or have two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T> SingleAsync<T>(this IAsyncQuery<T> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SingleAsync();
-			}
-
-			return Head(source, single: true, orDefault: false, default!);
-		}
-
-		/// <summary>Returns the first and only element of an async sequence, or an exception if it is empty or have two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T> SingleAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SingleAsync(predicate);
-			}
-
-			return Head(source.Where(predicate), single: true, orDefault: false, default!);
-		}
-
-		/// <summary>Returns the first and only element of an async sequence, or an exception if it is empty or have two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T> SingleAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SingleAsync(predicate);
-			}
-
-			return Head(source.Where(predicate), single: true, orDefault: false, default!);
-		}
-
-		#endregion
-
-		#region SingleOrDefaultAsync...
-
-		/// <summary>Returns the first and only element of an async sequence, the default value for the type if it is empty, or an exception if it has two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T?> SingleOrDefaultAsync<T>(this IAsyncQuery<T> source) => SingleOrDefaultAsync(source, default(T?));
-
-		/// <summary>Returns the first and only element of an async sequence, the default value for the type if it is empty, or an exception if it has two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T> SingleOrDefaultAsync<T>(this IAsyncQuery<T> source, T defaultValue)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SingleOrDefaultAsync(defaultValue);
-			}
-
-			return Head(source, single: true, orDefault: true, defaultValue);
-		}
-
-		/// <summary>Returns the first and only element of an async sequence, the default value for the type if it is empty, or an exception if it has two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T?> SingleOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate) => SingleOrDefaultAsync(source, predicate!, default(T?));
-
-		/// <summary>Returns the first and only element of an async sequence, the default value for the type if it is empty, or an exception if it has two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T> SingleOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate, T defaultValue)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SingleOrDefaultAsync(predicate, defaultValue);
-			}
-
-			return Head(source.Where(predicate), single: true, orDefault: true, defaultValue);
-		}
-
-		/// <summary>Returns the first and only element of an async sequence, the default value for the type if it is empty, or an exception if it has two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T?> SingleOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate) => SingleOrDefaultAsync(source, predicate!, default(T?));
-
-		/// <summary>Returns the first and only element of an async sequence, the default value for the type if it is empty, or an exception if it has two or more elements</summary>
-		/// <remarks>Will need to call MoveNext at least twice to ensure that there is no second element.</remarks>
-		public static Task<T> SingleOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate, T defaultValue)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SingleOrDefaultAsync(predicate, defaultValue);
-			}
-
-			return Head(source.Where(predicate), single: true, orDefault: true, defaultValue);
-		}
-
-		#endregion
-
-		#region LastAsync...
-
-		/// <summary>Returns the last element of an async sequence, or an exception if it is empty</summary>
-		public static Task<T> LastAsync<T>(this IAsyncQuery<T> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.LastAsync();
-			}
-
-			return Impl(source);
-
-			static async Task<T> Impl(IAsyncQuery<T> source)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				bool found = false;
-				T last = default!;
-
-				while(await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					found = true;
-					last = iterator.Current;
-				}
-
-				return found ? last : throw ErrorNoElements();
-			}
-		}
-
-		/// <summary>Returns the last element of an async sequence, or an exception if it is empty</summary>
-		public static Task<T> LastAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.LastAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, bool> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				T result = default!;
-				bool found = false;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var item = iterator.Current;
-					if (predicate(item))
-					{
-						found = true;
-						result = item;
-					}
-				}
-
-				return found ? result : throw ErrorNoElements();
-			}
-		}
-
-		/// <summary>Returns the last element of an async sequence, or an exception if it is empty</summary>
-		public static Task<T> LastAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.LastAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				T result = default!;
-				bool found = false;
-				var ct = source.Cancellation;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var item = iterator.Current;
-					if (await predicate(item, ct).ConfigureAwait(false))
-					{
-						found = true;
-						result = item;
-					}
-				}
-
-				return found ? result : throw ErrorNoElements();
-			}
-		}
-
-		#endregion
-
-		#region LastOrDefaultAsync...
-
-		/// <summary>Returns the last element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T?> LastOrDefaultAsync<T>(this IAsyncQuery<T> source) => LastOrDefaultAsync(source, default(T?));
-
-		/// <summary>Returns the last element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T> LastOrDefaultAsync<T>(this IAsyncQuery<T> source, T defaultValue)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.LastOrDefaultAsync(defaultValue);
-			}
-
-			return Impl(source, defaultValue);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, T defaultValue)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				if (!await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					return defaultValue;
-				}
-
-				T result;
-				do
-				{
-					result = iterator.Current;
-				}
-				while (await iterator.MoveNextAsync().ConfigureAwait(false));
-
-				return result;
-			}
-		}
-
-		/// <summary>Returns the last element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T?> LastOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate) => LastOrDefaultAsync(source, predicate!, default(T?));
-
-		/// <summary>Returns the last element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T> LastOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate, T defaultValue)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.LastOrDefaultAsync(predicate, defaultValue);
-			}
-
-			return Impl(source, predicate, defaultValue);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, bool> predicate, T defaultValue)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				T result = defaultValue;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var item = iterator.Current;
-					if (predicate(item))
-					{
-						result = iterator.Current;
-					}
-				}
-				return result;
-			}
-		}
-
-		/// <summary>Returns the last element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T?> LastOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate) => LastOrDefaultAsync(source, predicate!, default(T?));
-
-		/// <summary>Returns the last element of an async sequence, or the default value for the type if it is empty</summary>
-		public static Task<T> LastOrDefaultAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate, T defaultValue)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.LastOrDefaultAsync(predicate, defaultValue);
-			}
-
-			return Impl(source, predicate, defaultValue);
-
-			static async Task<T> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate, T defaultValue)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				T result = defaultValue;
-				var ct = source.Cancellation;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var item = iterator.Current;
-					if (await predicate(item, ct).ConfigureAwait(false))
-					{
-						result = iterator.Current;
-					}
-				}
-				return result;
-			}
-		}
-
-		#endregion
-
 		#region ElemetAtAsync...
 
 		/// <summary>Returns the element at a specific location of an async sequence, or an exception if there are not enough elements</summary>
@@ -1504,444 +920,6 @@ namespace SnowBank.Linq
 
 		#endregion
 
-		#region CountAsync...
-
-		/// <summary>Returns the number of elements in an async sequence.</summary>
-		public static Task<int> CountAsync<T>(this IAsyncQuery<T> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.CountAsync();
-			}
-
-			return Impl(source);
-
-			static async Task<int> Impl(IAsyncQuery<T> source)
-			{
-				int count = 0;
-				await ForEachAsync(source, (_) => { ++count; }).ConfigureAwait(false);
-				return count;
-			}
-		}
-
-		/// <summary>Returns a number that represents how many elements in the specified async sequence satisfy a condition.</summary>
-		public static Task<int> CountAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.CountAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<int> Impl(IAsyncQuery<T> source, Func<T, bool> predicate)
-			{
-				int count = 0;
-
-				await ForEachAsync(source, (x) =>
-				{
-					if (predicate(x))
-					{
-						checked { ++count; }
-					}
-				}).ConfigureAwait(false);
-
-				return count;
-			}
-		}
-
-		/// <summary>Returns a number that represents how many elements in the specified async sequence satisfy a condition.</summary>
-		public static Task<int> CountAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.CountAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<int> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate)
-			{
-				int count = 0;
-
-				await ForEachAsync(source, async (x, ct) => {
-					if (await predicate(x, ct).ConfigureAwait(false))
-					{
-						checked { ++count; }
-					}
-				}).ConfigureAwait(false);
-
-				return count;
-			}
-		}
-
-		#endregion
-
-		#region SumAsync...
-
-		/// <summary>Returns the sum of all elements in the specified async sequence that satisfy a condition.</summary>
-		public static Task<T> SumAsync<T>(this IAsyncQuery<T> source)
-			where T : INumberBase<T>
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.SumAsync();
-			}
-
-			if (typeof(T) == typeof(int)) return (Task<T>) (object) SumAsyncInt32Impl((IAsyncQuery<int>) source);
-			if (typeof(T) == typeof(long)) return (Task<T>) (object) SumAsyncInt64Impl((IAsyncQuery<long>) source);
-			if (typeof(T) == typeof(float)) return (Task<T>) (object) SumAsyncFloatImpl((IAsyncQuery<float>) source);
-			if (typeof(T) == typeof(double)) return (Task<T>) (object) SumAsyncDoubleImpl((IAsyncQuery<double>) source);
-			if (typeof(T) == typeof(decimal)) return (Task<T>) (object) SumAsyncDecimalImpl((IAsyncQuery<decimal>) source);
-
-			return SumAsyncImpl(source);
-		}
-
-		internal static async Task<T> SumAsyncImpl<T>(IAsyncQuery<T> source) where T : INumberBase<T>
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			T sum = T.Zero;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				sum = checked(sum + iterator.Current);
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence that satisfy a condition.</summary>
-		public static Task<T?> SumAsync<T>(this IAsyncQuery<T?> source)
-			where T : struct, INumberBase<T>
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T?> query)
-			{
-				return query.SumAsync();
-			}
-
-			if (typeof(T) == typeof(int)) return (Task<T?>) (object) SumAsyncInt32Impl((IAsyncQuery<int?>) source);
-			if (typeof(T) == typeof(long)) return (Task<T?>) (object) SumAsyncInt64Impl((IAsyncQuery<long?>) source);
-			if (typeof(T) == typeof(float)) return (Task<T?>) (object) SumAsyncFloatImpl((IAsyncQuery<float?>) source);
-			if (typeof(T) == typeof(double)) return (Task<T?>) (object) SumAsyncDoubleImpl((IAsyncQuery<double?>) source);
-			if (typeof(T) == typeof(decimal)) return (Task<T?>) (object) SumAsyncDecimalImpl((IAsyncQuery<decimal?>) source);
-
-			return SumAsyncNullableImpl(source);
-		}
-
-		internal static async Task<T?> SumAsyncNullableImpl<T>(IAsyncQuery<T?> source) where T : struct, INumberBase<T>
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			T sum = T.Zero;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				var item = iterator.Current;
-				if (item is not null)
-				{
-					sum = checked(sum + item.GetValueOrDefault());
-				}
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<int> SumAsync(this IAsyncQuery<int> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<int> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncInt32Impl(source);
-		}
-
-		internal static async Task<int> SumAsyncInt32Impl(IAsyncQuery<int> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			int sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				checked { sum += iterator.Current; }
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<int?> SumAsync(this IAsyncQuery<int?> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<int?> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncInt32Impl(source);
-		}
-
-		internal static async Task<int?> SumAsyncInt32Impl(IAsyncQuery<int?> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			int sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				var item = iterator.Current;
-				if (item is not null)
-				{
-					sum = checked(sum + item.GetValueOrDefault());
-				}
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<long> SumAsync(this IAsyncQuery<long> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<long> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncInt64Impl(source);
-		}
-
-		internal static async Task<long> SumAsyncInt64Impl(IAsyncQuery<long> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			long sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				checked { sum += iterator.Current; }
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<long?> SumAsync(this IAsyncQuery<long?> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<long?> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncInt64Impl(source);
-		}
-
-		internal static async Task<long?> SumAsyncInt64Impl(IAsyncQuery<long?> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			long sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				var item = iterator.Current;
-				if (item is not null)
-				{
-					sum = checked(sum + item.GetValueOrDefault());
-				}
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<float> SumAsync(this IAsyncQuery<float> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<float> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncFloatImpl(source);
-		}
-
-		internal static async Task<float> SumAsyncFloatImpl(IAsyncQuery<float> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			//note: like Enumerable and AsyncEnumerable, we will also use a double as the accumulator (to reduce precision loss)
-
-			double sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				sum += iterator.Current;
-			}
-
-			return (float) sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<float?> SumAsync(this IAsyncQuery<float?> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<float?> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncFloatImpl(source);
-		}
-
-		internal static async Task<float?> SumAsyncFloatImpl(IAsyncQuery<float?> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			//note: like Enumerable and AsyncEnumerable, we will also use a double as the accumulator (to reduce precision loss)
-			double sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				var item = iterator.Current;
-				if (item is not null)
-				{
-					sum += item.GetValueOrDefault();
-				}
-			}
-
-			return (float) sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<double> SumAsync(this IAsyncQuery<double> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<double> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncDoubleImpl(source);
-		}
-
-		internal static async Task<double> SumAsyncDoubleImpl(IAsyncQuery<double> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			double sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				sum += iterator.Current;
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<double?> SumAsync(this IAsyncQuery<double?> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<double?> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncDoubleImpl(source);
-		}
-
-		internal static async Task<double?> SumAsyncDoubleImpl(IAsyncQuery<double?> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			double sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				var item = iterator.Current;
-				if (item is not null)
-				{
-					sum += item.GetValueOrDefault();
-				}
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<decimal> SumAsync(this IAsyncQuery<decimal> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<decimal> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncDecimalImpl(source);
-		}
-
-		internal static async Task<decimal> SumAsyncDecimalImpl(IAsyncQuery<decimal> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			decimal sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				sum += iterator.Current;
-			}
-
-			return sum;
-		}
-
-		/// <summary>Returns the sum of all elements in the specified async sequence.</summary>
-		public static Task<decimal?> SumAsync(this IAsyncQuery<decimal?> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<decimal?> query)
-			{
-				return query.SumAsync();
-			}
-
-			return SumAsyncDecimalImpl(source);
-		}
-
-		internal static async Task<decimal?> SumAsyncDecimalImpl(IAsyncQuery<decimal?> source)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			decimal sum = 0;
-			while (await iterator.MoveNextAsync().ConfigureAwait(false))
-			{
-				var item = iterator.Current;
-				if (item is not null)
-				{
-					sum += item.GetValueOrDefault();
-				}
-			}
-
-			return sum;
-		}
-
-		#endregion
-
 		#region AsEnumerable...
 
 		/// <summary>Helper method that checks that the cancellation token is the same as the source</summary>
@@ -1955,384 +933,6 @@ namespace SnowBank.Linq
 				throw new InvalidOperationException("The CancellationToken that is passed to GetAsyncEnumerator() MUST be the same as the source!");
 			}
 			return query.GetAsyncEnumerator(hint);
-		}
-
-		/// <summary>Adapts this query into the equivalent <see cref="IAsyncEnumerable{T}"/></summary>
-		/// <param name="source">Source query that will be adapted into an <see cref="IAsyncEnumerable{T}"/></param>
-		/// <param name="hint">Hint passed to the source provider.</param>
-		/// <returns>Sequence that will asynchronously return the results of this query.</returns>
-		/// <remarks>
-		/// <para>For best performance, the caller should take care to provide a <see cref="hint"/> that matches how this query will be consumed downstream.</para>
-		/// <para>If the hint does not match, performance may be degraded.
-		/// For example, if the caller will consumer this query using <c>await foreach</c> or <c>ToListAsync</c>, but uses <see cref="AsyncIterationHint.Iterator"/>, the provider may fetch small pages initially, before ramping up.
-		/// The opposite is also true if the caller uses <see cref="AsyncIterationHint.All"/> but consumes the query using <c>AnyAsync()</c> or <c>FirstOrDefaultAsync</c>, the provider may fetch large pages and waste most of it except the first few elements.
-		/// </para>
-		/// </remarks>
-		public static IAsyncEnumerable<T> ToAsyncEnumerable<T>(this IAsyncQuery<T> source, AsyncIterationHint hint = AsyncIterationHint.Default)
-			=> source is IAsyncLinqQuery<T> query ? query.ToAsyncEnumerable() : new ConfiguredAsyncEnumerable<T>(source, hint);
-
-		public static IAsyncEnumerable<T> WantAll<T>(this IAsyncQuery<T> source)
-			=> new ConfiguredAsyncEnumerable<T>(source, AsyncIterationHint.All);
-
-		/// <summary>Exposes an async query as a regular <see cref="IAsyncEnumerable{T}"/>, with en explicit <see cref="AsyncIterationHint"/></summary>
-		internal sealed class ConfiguredAsyncEnumerable<T> : IAsyncEnumerable<T>
-		{
-
-			private IAsyncQuery<T> Source { get; }
-
-			private AsyncIterationHint Hint { get; }
-
-			public ConfiguredAsyncEnumerable(IAsyncQuery<T> source, AsyncIterationHint hint)
-			{
-				this.Source = source;
-				this.Hint = hint;
-			}
-
-			public CancellationToken Cancellation => this.Source.Cancellation;
-
-			[MustDisposeResource]
-			public IAsyncEnumerator<T> GetAsyncEnumerator(CancellationToken ct = default)
-			{
-				//BUGBUG: if ct is not None and not the same as the source, we should maybe mix them!?
-				return this.Source.GetAsyncEnumerator(this.Hint);
-			}
-
-		}
-
-		public static IAsyncLinqQuery<T> ToAsyncQuery<T>(this IAsyncEnumerable<T> source, CancellationToken ct = default)
-		{
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				if (ct.CanBeCanceled && !ct.Equals(query.Cancellation))
-				{
-					throw new InvalidOperationException("The CancellationToken that is passed to ToAsyncQuery() MUST be the same as the source!");
-				}
-				return query;
-			}
-			return new AdapterAsyncEnumerable<T>(source, ct);
-		}
-
-		internal sealed class AdapterAsyncEnumerable<T> : AsyncLinqIterator<T>
-		{
-
-			private IAsyncEnumerable<T> Source { get; }
-
-			private IAsyncEnumerator<T>? Iterator { get; set; }
-
-			public AdapterAsyncEnumerable(IAsyncEnumerable<T> source, CancellationToken ct)
-			{
-				this.Source = source;
-				this.Cancellation = ct;
-			}
-
-			public override CancellationToken Cancellation { get; }
-
-			/// <inheritdoc />
-			protected override AdapterAsyncEnumerable<T> Clone() => new(this.Source, this.Cancellation);
-
-			/// <inheritdoc />
-			protected override ValueTask<bool> OnFirstAsync()
-			{
-				this.Iterator = this.Source.GetAsyncEnumerator(this.Cancellation);
-				return new(true);
-			}
-
-			/// <inheritdoc />
-			protected override async ValueTask<bool> OnNextAsync()
-			{
-				var iterator = this.Iterator;
-				if (iterator == null || !(await iterator.MoveNextAsync().ConfigureAwait(false)))
-				{
-					return await this.Completed().ConfigureAwait(false);
-				}
-				return Publish(iterator.Current);
-			}
-
-			/// <inheritdoc />
-			protected override async ValueTask Cleanup()
-			{
-				var iterator = this.Iterator;
-				this.Iterator = null;
-				if (iterator != null)
-				{
-					await iterator.DisposeAsync().ConfigureAwait(false);
-				}
-			}
-
-#if NET10_0_OR_GREATER
-
-			/// <inheritdoc />
-			public override Task<T[]> ToArrayAsync() => this.Source.ToArrayAsync(this.Cancellation).AsTask();
-
-			/// <inheritdoc />
-			public override Task<List<T>> ToListAsync() => this.Source.ToListAsync(this.Cancellation).AsTask();
-
-			/// <inheritdoc />
-			public override Task<bool> AnyAsync() => this.Source.AnyAsync(this.Cancellation).AsTask();
-
-			/// <inheritdoc />
-			public override Task<bool> AnyAsync(Func<T, bool> predicate) => this.Source.AnyAsync(predicate, this.Cancellation).AsTask();
-
-			/// <inheritdoc />
-			public override IAsyncLinqQuery<TNew> Select<TNew>(Func<T, TNew> selector) => new AdapterAsyncEnumerable<TNew>(this.Source.Select(selector), this.Cancellation);
-
-			/// <inheritdoc />
-			public override IAsyncLinqQuery<TNew> Select<TNew>(Func<T, int, TNew> selector) => new AdapterAsyncEnumerable<TNew>(this.Source.Select(selector), this.Cancellation);
-
-			/// <inheritdoc />
-			public override IAsyncLinqQuery<T> Where(Func<T, bool> predicate) => new AdapterAsyncEnumerable<T>(this.Source.Where(predicate), this.Cancellation);
-
-			/// <inheritdoc />
-			public override IAsyncLinqQuery<T> Where(Func<T, int, bool> predicate) => new AdapterAsyncEnumerable<T>(this.Source.Where(predicate), this.Cancellation);
-
-#endif
-
-		}
-
-		#endregion
-
-		#region Min/Max...
-
-		/// <summary>Returns the smallest value in the specified async sequence</summary>
-		public static Task<T?> MinAsync<T>(this IAsyncQuery<T> source, IComparer<T>? comparer = null)
-		{
-			Contract.NotNull(source);
-			comparer ??= Comparer<T>.Default;
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.MinAsync(comparer);
-			}
-
-			return MinAsyncImpl(source, comparer);
-
-		}
-
-		internal static async Task<T?> MinAsyncImpl<T>(IAsyncQuery<T> source, IComparer<T> comparer)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			if (default(T) is null)
-			{
-				// we will return null if the query is empty, or only contains null
-				var min = default(T);
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var candidate = iterator.Current;
-					if (candidate is not null && (min is null || comparer.Compare(candidate, candidate) > 0))
-					{
-						min = candidate;
-					}
-				}
-				return min;
-			}
-			else
-			{
-				// we will throw if the query is empty
-
-				// get the first
-				if (!await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					throw ErrorNoElements();
-				}
-				var min = iterator.Current;
-
-				// compare with the rest
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var candidate = iterator.Current;
-					if (comparer.Compare(candidate, min) < 0)
-					{
-						min = candidate;
-					}
-				}
-
-				return min;
-			}
-		}
-
-		/// <summary>Returns the largest value in the specified async sequence</summary>
-		public static Task<T?> MaxAsync<T>(this IAsyncQuery<T> source, IComparer<T>? comparer = null)
-		{
-			Contract.NotNull(source);
-			comparer ??= Comparer<T>.Default;
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.MaxAsync(comparer);
-			}
-
-			return MaxAsyncImpl(source, comparer);
-		}
-
-		internal static async Task<T?> MaxAsyncImpl<T>(IAsyncQuery<T> source, IComparer<T> comparer)
-		{
-			await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-			if (default(T) is null)
-			{
-				var max = default(T);
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var candidate = iterator.Current;
-					if (candidate is not null && (max is null || comparer.Compare(candidate, candidate) > 0))
-					{
-						max = candidate;
-					}
-				}
-				return max;
-			}
-			else
-			{
-				// get the first
-				if (!await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					throw ErrorNoElements();
-				}
-				var max = iterator.Current;
-
-				// compare with the rest
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					var candidate = iterator.Current;
-					if (comparer.Compare(candidate, max) > 0)
-					{
-						max = candidate;
-					}
-				}
-
-				return max;
-			}
-		}
-
-		#endregion
-
-		#region Any/None...
-
-		/// <summary>Determines whether an async sequence contains any elements.</summary>
-		/// <remarks>This is the logical equivalent to "source.Count() > 0" but can be better optimized by some providers</remarks>
-		public static Task<bool> AnyAsync<T>(this IAsyncQuery<T> source)
-		{
-			Contract.NotNull(source);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.AnyAsync();
-			}
-
-			return Impl(source);
-
-			static async Task<bool> Impl(IAsyncQuery<T> source)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Head);
-				return await iterator.MoveNextAsync().ConfigureAwait(false);
-			}
-		}
-
-		/// <summary>Determines whether any element of an async sequence satisfies a condition.</summary>
-		public static Task<bool> AnyAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.AnyAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<bool> Impl(IAsyncQuery<T> source, Func<T, bool> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Iterator);
-
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (predicate(iterator.Current)) return true;
-				}
-
-				return false;
-			}
-		}
-
-		/// <summary>Determines whether any element of an async sequence satisfies a condition.</summary>
-		public static Task<bool> AnyAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.AnyAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<bool> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.Iterator);
-
-				var ct = source.Cancellation;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (await predicate(iterator.Current, ct).ConfigureAwait(false)) return true;
-				}
-
-				return false;
-			}
-		}
-
-		/// <summary>Determines whether any element of an async sequence satisfies a condition.</summary>
-		public static Task<bool> AllAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, bool> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.AllAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<bool> Impl(IAsyncQuery<T> source, Func<T, bool> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (!predicate(iterator.Current)) return false;
-				}
-
-				return true;
-			}
-		}
-
-		/// <summary>Determines whether any element of an async sequence satisfies a condition.</summary>
-		public static Task<bool> AllAsync<T>(this IAsyncQuery<T> source, [InstantHandle] Func<T, CancellationToken, Task<bool>> predicate)
-		{
-			Contract.NotNull(source);
-			Contract.NotNull(predicate);
-
-			if (source is IAsyncLinqQuery<T> query)
-			{
-				return query.AllAsync(predicate);
-			}
-
-			return Impl(source, predicate);
-
-			static async Task<bool> Impl(IAsyncQuery<T> source, Func<T, CancellationToken, Task<bool>> predicate)
-			{
-				await using var iterator = source.GetAsyncEnumerator(AsyncIterationHint.All);
-
-				var ct = source.Cancellation;
-				while (await iterator.MoveNextAsync().ConfigureAwait(false))
-				{
-					if (!(await predicate(iterator.Current, ct).ConfigureAwait(false))) return false;
-				}
-
-				return true;
-			}
 		}
 
 		#endregion
